requirements.txt
fastapi
langchain
openai
uvicorn
python-dotenv
pydantic
streamlit
requests
pypdf